+++ central/detection/service/singleton.go
@@ -7,6 +7,7 @@ import (
 	"github.com/stackrox/rox/central/detection/deploytime"
 	"github.com/stackrox/rox/central/enrichment"
 	imageDatastore "github.com/stackrox/rox/central/image/datastore"
+	"github.com/stackrox/rox/central/notifier/processor"
 	"github.com/stackrox/rox/central/risk/manager"
 	"github.com/stackrox/rox/pkg/sync"
 )
@@ -24,6 +25,7 @@ func initialize() {
 		cveDataStore.Singleton(),
 		enrichment.Singleton(),
 		buildTimeDetection.SingletonDetector(),
+		processor.Singleton(),
 		deploytime.SingletonDetector(),
 		deploytime.SingletonPolicySet())
 }